# public/swagger/qslp_openapi3.yaml
openapi: 3.0.1
info:
  title: ASLP API
  description: |
    API Documentation for the ASLP v1 & v2 API
    <h4>API Endpoint URL</h4>
    The endpoint for our REST API is:  <code>https://aslp.explorer.sh/api</code>
    <br /><br />
        <h4>Master Address</h4>
    The Master Address for ASLP transactions is:  <code>ARKQXzHvEWXgfCgAcJWJQKUMus5uE6Yckr</code>
    <br /><br />
    <h4>Datetime Formatting</h4>
    Unless otherwise stated, all timestamps are returned in ISO8601 format in UTC. Example: "2018-04-01T00:00:00Z" or "2018-04-01T00:00:00.000Z".   If a timestamp is returned as an Integer, then that is a Unix timestamp.
    <br /><br />
    <h4>Number Formatting</h4>
    All finance data such as price, quantity, and fee should be arbitrary precision numbers and string representation. Example: "5.39300000"
    <br /><br />
    <h4>Error Codes</h4>
    <table class="table table-striped">
    <thead>
    <tr><td>Error Code</td><td>HTTP Status Code</td><td>Message</td><td>Details</td></tr>
    </thead>
    <tbody>
    <tr>
    <td>400</td>
    <td>400</td>
    <td style="text-align: left">Unknown error</td>
    <td style="text-align: left">An unknown error occurred somewhere in the system</td>
    </tr>
    <tr>
    <td>402</td>
    <td>400</td>
    <td style="text-align: left">Method not found</td>
    <td style="text-align: left">The requested API method was not found</td>
    </tr>
    <tr>
    <td>403</td>
    <td>401</td>
...
    <td style="text-align: left"></td>
    </tr>
...
servers:
...
paths:
...
  /token/{id}:
    get:
      tags:
        - ASLP1/ASLP2
      summary: Get token info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Token ID Hex
      responses:
        200:
          description: Token Information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        400:
          description: Currency not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /transaction/{txid}:
    get:
      tags:
        - ASLP1/ASLP2
      summary: Get a listing of tokens
      parameters:
        - name: txid
          in: path
          required: true
          schema:
            type: string
          description: Transaction ID
      responses:
        200:
          description: Transaction Information
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TransactionASLP1"
                  - $ref: "#/components/schemas/TransactionASLP2"
        400:
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /transactions:
    get:
      tags:
        - ASLP1/ASLP2
      summary: Get a listing of transactions
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: Transaction List
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/TransactionASLP1"
                    - $ref: "#/components/schemas/TransactionASLP2"
        400:
          description: Tokens not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Token:
      type: object
      properties:
        schema_version:
          type: integer
          description: Schema Version
        type:
          type: string
          description: ASLP1 or ASLP2
        paused:
          type: boolean
          description: token paused for activity
        tokenDetails:
          type: object
          properties:
            schema_version:
              type: integer
            ownerAddress:
              type: string
              description: Ark address of token contract controller
            tokenIdHex:
              type: string
              description: Token ID Code
            versionType:
              type: integer
              description: ASLP version (1 or 2)
            genesis_timestamp:
              type: string
              format: date-time
              description: Datetime token was created
            genesis_timestamp_unix:
              type: integer
              description: Unix timestamp token was created
            symbol:
              type: string
              description: Ticker code for token
            name:
              type: string
              description: Token name
            documentUri:
              type: string
              description: Generally a url pointing to token homepage
            decimals:
              type: integer
              description: Number of decimal places this token has
            genesisQuantity:
              type: string
              description: String formatted number of the genesis quantity in raw form (no decimals)
            pausable:
              type: boolean
              description: Is the token pausable
            mintable:
              type: boolean
              description: Can additional tokens be minted
        tokenStats:
          type: object
          properties:
            schema_version:
              type: integer
              description: Schema Version Number
            block_created_height:
              type: integer
              description: Block height the token was created at
            block_created_id:
              type: string
              description: Block ID/Hash the token was created in
            block_last_active_send:
              type: integer
              description: Block height of last token transfer
            block_last_active_mint:
              type: integer
              description: Block height of last mint action
            creation_transaction_id:
              type: string
              description: Transaction ID the token was created with
            qty_valid_txns_since_genesis:
              type: integer
              description: Number of transactions on network since token created
            qty_valid_token_addresses:
              type: integer
              description: Number of addresses that have made transactions with this token
            qty_token_minted:
              type: string
              description: String formatted number of total tokens minted in raw form (no decimals)
            qty_token_burned:
              type: string
              description: String formatted number of total tokens burned in raw form (no decimals)
            qty_token_circulating_supply:
              type: string
              description: String formatted number of total tokens in circulation in raw form (no decimals)
            qty_ark_spent:
              type: string
              description: String formatted number of total ARK spend for all transactions in raw form (no decimals)
        lastUpdatedBlock:
          type: integer
          description: Last block height a transaction has occurred for this token
...
    TransactionASLP1:
      type: object
      properties:
        schema_version:
          type: integer
        txid:
          type: string
        blockId:
          type: string
        blockHeight:
          type: integer
        valid:
          type: boolean
        invalidReason:
          type: string
        transactionDetails:
          type: object
          properties:
            schema:
              $ref: "#/components/schemas/TransactionDetailsASLP1"
    TransactionDetailsASLP1:
      type: object
      properties:
        schema_version:
          type: integer
        transactionType:
          type: string
        senderAddress:
          type: string
        tokenIdHex:
          type: string
        versionType:
          type: integer
        timestamp:
          type: string
          format: date-time
        timestamp_unix:
          type: integer
        symbol:
          type: string
        name:
          type: string
        documentUri:
          type: string
        decimals:
          type: integer
        genesisOrMintQuantity:
          type: string
        sendOutput:
          type: object
          properties:
            schema:
              $ref: "#/components/schemas/TransactionOutputASLP1"
        note:
          type: string
        amount_ark:
          type: string
        fee_ark:
          type: string
    TransactionOutputASLP1:
      type: object
      properties:
        schema_version:
          type: integer
        address:
          type: string
        amount:
          type: string
    TransactionASLP2:
      type: object
      properties:
        schema_version:
          type: integer
        txid:
          type: string
        blockId:
          type: string
        blockHeight:
          type: integer
        valid:
          type: boolean
        invalidReason:
          type: string
        transactionDetails:
          type: object
          properties:
            schema:
              $ref: "#/components/schemas/TransactionDetailsASLP2"
    TransactionDetailsASLP2:
      type: object
      properties:
        schema_version:
          type: integer
        transactionType:
          type: string
        senderAddress:
          type: string
        tokenIdHex:
          type: string
        versionType:
          type: integer
        timestamp:
          type: string
          format: date-time
        timestamp_unix:
          type: integer
        symbol:
          type: string
        name:
          type: string
        documentUri:
          type: string
        note:
          type: string
        amount_ark:
          type: string
        fee_ark:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              format: int32
            message:
              type: string
              enum:
                - Internal Server Error
                - Gateway Timeout
                - Service Unavailable
                - Symbol not found
                - Authorisation required
                - Authorisation failed
                - Validation error
                - Insufficient funds
            description:
              type: string